Added third level
Since pieces two and three are both vertical pipes I figured that I could cheat and use the same tile map. This works but it did not give a nice visual indication to the user that the had migrated between two pieces. So I added a new tile map which overlays the original tile map to add differences to the scene.