The business logic portion of the game is now compiling and passing all tests
Unfortunately I needed to duplicate some of the implementation for Label Retriever, so that I could create my own version of a Label Retriever versus always recieving the base implementation. While I am not a fan of duplicating code I am less interested in trying to open up every aspect of this class as public or protected. Mainly because that would require two NuGet package updates prior to me being able to even test out the code changes. This method works, and therefore is the one that I will go with.

I also removed a bunch of unit tests from the Label Retriever. I do not see a reason to retest that I can get a specific label since the base NuGet package is already testing that. So I should consider this functionality out-of-the-box so to speak and not have to validate it. Therefore I am only exercising my new implementations of GetLabelRetriever and Reset.