Business Logic publishing all dependencies on build
Since Unity does not want to support NuGet I needed to setup a way to push all of the assemblies that the business logic is dependent upon to the references folder that I added to Unity. Through some searching I found the command `dotnet publish`, see: https://docs.microsoft.com/en-us/dotnet/core/tools/dotnet-publish, and this enables me to specify an output directory and push all of the assemblies required for a specific project. Once I tested this locally I updated the post-build events so that this would happen automatically.

This is a great addition to this development pipeline as I have added NuGet support for the business logic project, which enables me to quickly update to a later version of a particular dll. Using this automation will prevent me from being in `dll hell` where I will need to match and double check specific versions. In addition this automation will prevent me from having to manually manage all of these dependencies.

Finally I updated the existing Unity logic to reference the code in the dependent NuGet files for the state machine.